+++ modules/core/02-client/keeper/migrations_test.go
@@ -2,45 +2,9 @@ package keeper_test
 
 import (
 	"github.com/cosmos/ibc-go/v10/modules/core/02-client/keeper"
-	"github.com/cosmos/ibc-go/v10/modules/core/02-client/types"
 	host "github.com/cosmos/ibc-go/v10/modules/core/24-host"
 	ibcexported "github.com/cosmos/ibc-go/v10/modules/core/exported"
 )
-
-// TestMigrateParams tests the migration for the client params
-func (suite *KeeperTestSuite) TestMigrateParams() {
-	testCases := []struct {
-		name           string
-		malleate       func()
-		expectedParams types.Params
-	}{
-		{
-			"success: default params",
-			func() {
-				params := types.DefaultParams()
-				subspace := suite.chainA.GetSimApp().GetSubspace(ibcexported.ModuleName)
-				subspace.SetParamSet(suite.chainA.GetContext(), &params)
-			},
-			types.DefaultParams(),
-		},
-	}
-
-	for _, tc := range testCases {
-		suite.Run(tc.name, func() {
-			suite.SetupTest() // reset
-
-			tc.malleate()
-
-			ctx := suite.chainA.GetContext()
-			migrator := keeper.NewMigrator(suite.chainA.GetSimApp().IBCKeeper.ClientKeeper)
-			err := migrator.MigrateParams(ctx)
-			suite.Require().NoError(err)
-
-			params := suite.chainA.GetSimApp().IBCKeeper.ClientKeeper.GetParams(ctx)
-			suite.Require().Equal(tc.expectedParams, params)
-		})
-	}
-}
 
 func (suite *KeeperTestSuite) TestMigrateToStatelessLocalhost() {
 	// set localhost in state