+++ modules/core/02-client/types/codec_test.go
@@ -192,11 +192,6 @@ func (suite *TypesTestSuite) TestCodecTypeRegistration() {
 			nil,
 		},
 		{
-			"success: MsgSubmitMisbehaviour",
-			sdk.MsgTypeURL(&types.MsgSubmitMisbehaviour{}),
-			nil,
-		},
-		{
 			"success: MsgRecoverClient",
 			sdk.MsgTypeURL(&types.MsgRecoverClient{}),
 			nil,