+++ modules/core/02-client/types/expected_keepers.go
@@ -3,10 +3,7 @@ package types
 import (
 	"context"
 
-	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
+	upgradetypes "cosmossdk.io/x/upgrade/types"
 )
 
 // UpgradeKeeper expected upgrade keeper
@@ -17,9 +14,4 @@ type UpgradeKeeper interface {
 	GetUpgradedConsensusState(ctx context.Context, lastHeight int64) ([]byte, error)
 	SetUpgradedConsensusState(ctx context.Context, planHeight int64, bz []byte) error
 	ScheduleUpgrade(ctx context.Context, plan upgradetypes.Plan) error
-}
-
-// ParamSubspace defines the expected Subspace interface for module parameters.
-type ParamSubspace interface {
-	GetParamSet(ctx sdk.Context, ps paramtypes.ParamSet)
 }