+++ modules/core/keeper/msg_server.go
@@ -121,24 +121,6 @@ func (k *Keeper) UpgradeClient(goCtx context.Context, msg *clienttypes.MsgUpgrad
 	return &clienttypes.MsgUpgradeClientResponse{}, nil
 }
 
-// SubmitMisbehaviour defines a rpc handler method for MsgSubmitMisbehaviour.
-// Warning: DEPRECATED
-// This handler is redundant as `MsgUpdateClient` is now capable of handling both a Header and a Misbehaviour
-func (k *Keeper) SubmitMisbehaviour(goCtx context.Context, msg *clienttypes.MsgSubmitMisbehaviour) (*clienttypes.MsgSubmitMisbehaviourResponse, error) { //nolint:staticcheck // for now, we're using msgsubmitmisbehaviour.
-	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	misbehaviour, err := clienttypes.UnpackClientMessage(msg.Misbehaviour)
-	if err != nil {
-		return nil, err
-	}
-
-	if err = k.ClientKeeper.UpdateClient(ctx, msg.ClientId, misbehaviour); err != nil {
-		return nil, err
-	}
-
-	return &clienttypes.MsgSubmitMisbehaviourResponse{}, nil
-}
-
 // RecoverClient defines a rpc handler method for MsgRecoverClient.
 func (k *Keeper) RecoverClient(goCtx context.Context, msg *clienttypes.MsgRecoverClient) (*clienttypes.MsgRecoverClientResponse, error) {
 	if k.GetAuthority() != msg.Signer {