+++ CHANGELOG.md
@@ -43,9 +43,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Dependencies
 
 * [\#8451](https://github.com/cosmos/ibc-go/pull/8451) Bump **go** to **1.24**
+* (light-clients/08-wasm)[\#8500](https://github.com/cosmos/ibc-go/pull/8500) Bump **github.com/prysmaticlabs/prysm/v5** to **github.com/OffchainLabs/prysm/v6@v6.0.4**
 
 ### API Breaking
 * (apps) [\#8476](https://github.com/cosmos/ibc-go/pull/8476) Remove `ParamSubspace` from all `Keeper` constructors
+* (light-clients/08-wasm) [\#8511](https://github.com/cosmos/ibc-go/pull/8511) Remove deprecated `Checksums` type
+* (core/02-client) [\#8516](https://github.com/cosmos/ibc-go/pull/8516) Remove deprecated `SubmitMisbehaviour` message handler
 
 ### State Machine Breaking